requirements.txt
# Production dependencies
pydantic
pydantic-settings
fastapi
uvicorn
prometheus_client
requests
PyJWT
opensearch-py